+++ tests/Models/AnnotatedLargeExample.php
@@ -65,21 +65,9 @@ use Illuminate\Database\Eloquent\Model;
  * Apparently there's an issue with really long docblocks, they say!
  * Apparently there's an issue with really long docblocks, they say!
  * Apparently there's an issue with really long docblocks, they say!
- * Apparently there's an issue with really long docblocks, they say!
- * Apparently there's an issue with really long docblocks, they say!
- * Apparently there's an issue with really long docblocks, they say!
- * Apparently there's an issue with really long docblocks, they say!
- * Apparently there's an issue with really long docblocks, they say!
- * Apparently there's an issue with really long docblocks, they say!
- * Apparently there's an issue with really long docblocks, they say!
- * Apparently there's an issue with really long docblocks, they say!
- * Apparently there's an issue with really long docblocks, they say!
- * Apparently there's an issue with really long docblocks, they say!
- *
  *
  * @property \BenSampo\Enum\Tests\Enums\UserType|null $user_type
  * @see https://github.com/BenSampo/laravel-enum/issues/120
- *
  */
 class AnnotatedLargeExample extends Model
 {